RAG project/requirements.txt
langchain==0.0.37
huggingface-hub==0.13.4
sentence-transformers==2.2.0
transformers==4.11.3
faiss-cpu==1.7.1
pandas==1.3.5
python-dotenv==0.19.1
torch
python-docx